Update navigation and layout for improved user management
The changes in this commit involve updating navigation by switching the 'Administration' link to the 'Users' link. This was changed to help users more intuitively navigate to the user management ('GestionUsuarios') section of the site.

The Layout.cshtml was altered, specifically the divs and mains encapsulating the RenderBody() were rearranged for improved UI structure.

Finally, changes to the Index.cshtml in the 'GestionUsuarios' view have been committed to refine the user management interface. The alterations made provide a clearer, more optimised layout for displaying and managing users. These changes improve readability and ease of use for managing users.