Improve customer creation view design
A revamp in the visual design of the Create Customer view has been done aiming to facilitate user interaction. Changes include the addition of placeholder text in inputs, localization of labels and titles, and restructuration of the form for a better user experience. Moreover, the addition of icons to action buttons and restructuring the positioning and appearance of input fields brings clarity and aesthetic appeal to the page.