"Remove sensitive code and refactor permissions"
In this commit, sensitive data was removed from the codebase and admin level permissions refactored in the 'HuespedesController.cs'. The 'Database.cs' file was adjusted to take out the sensitive connection details that was posing a security risk. Database schema was also modified by removing the create statements for 'CheckIn' and 'CheckOut' tables. Authorization attributes have been added to 'Delete' and 'DeleteConfirmed' methods in 'HuespedesController.cs' to ensure they can only be accessed by an admin, therefore enhancing the system's security.