Refactor client search functionality into service
Refactoring was done on the client search functionality in the ClientesController, migrating the logic to a new service called ClienteService. This change ensures better organization and separation of concerns, making the code more maintainable. Additionally, minor changes have been made like changing "Reservas" to "ReservasController" in ReservasController and adding the ClienteService to the dependency injection container in the Program.cs file.