Refactor reservation process and rename associated views
This commit revises the reservation workflow. It renames the hotel reservation view from 'Cliente' (client) to 'Reservar' (reserve) and introduces a more direct reservation process. The 'Reservar' method in the 'Reservas' controller now starts a new reservation and redirects to a new 'ReservarHabitacion' (ReserveRoom) view. In the client list view, instead of a "select" button, there is now a "reserve" button that starts the reservation process.